Democratize parse_args!!
